refactor: move error_snapshotter.ts and error_tracker.ts to core package
