+++ packages/core/src/crawlers/index.ts
@@ -2,3 +2,5 @@ export * from './crawler_commons';
 export * from './crawler_extension';
 export * from './crawler_utils';
 export * from './statistics';
+export * from './error_tracker';
+export * from './error_snapshotter';
