+++ packages/utils/src/index.ts
@@ -7,8 +7,6 @@ export * from './internals/memory-info';
 export * from './internals/debug';
 export * as social from './internals/social';
 export * from './internals/typedefs';
-export * from './internals/error_tracker';
-export * from './internals/error_snapshotter';
 export * from './internals/open_graph_parser';
 export * from './internals/gotScraping';
 export * from './internals/robots';
