+++ test/core/error_tracker.test.ts
@@ -1,7 +1,5 @@
 /* eslint-disable no-multi-spaces */
-import exp from 'node:constants';
-
-import { ErrorTracker } from '../../packages/utils/src/internals/error_tracker';
+import { ErrorTracker } from '../../packages/core/src/crawlers/error_tracker';
 
 const random = () => Math.random().toString(36).slice(2);
